+++ ExpressoApp/components/ChecklistTask.js
