+++ ExpressoApp/components/QuantityInput.js
@@ -47,7 +47,7 @@ const QuantityInput = (props) => {
     <View style={styles.quantityRow}>
       <TouchableOpacity onPress={() => handleMinus()}>
         <Image
-          source={require('../../assets/minusButton.png')}
+          source={require('../assets/minusButton.png')}
           style={styles.quantityIcon}
         />
       </TouchableOpacity>
@@ -58,7 +58,7 @@ const QuantityInput = (props) => {
       />
       <TouchableOpacity onPress={() => handlePlus()}>
         <Image
-          source={require('../../assets/addButton.png')}
+          source={require('../assets/addButton.png')}
           style={styles.quantityIcon}
         />
       </TouchableOpacity>
