+++ ExpressoApp/screens/cart/CartScreen.js
@@ -7,8 +7,9 @@ import {
   TouchableOpacity, Image,
 } from 'react-native';
 import Cart from "./Cart";
-import AddToCartButton from "./AddToCartButton";
-import PaymentButton from "./PaymentButton";
+import AddToCartButton from "../../components/AddToCartButton";
+import PaymentButton from "../../components/PaymentButton";
+import Header from "../../components/Header";
 /**
  *
  * @return {JSX.Element}
@@ -23,15 +24,7 @@ const CartScreen = ({navigation}) => {
 
   return (
     <ScrollView>
-      <View style={styles.navBar}>
-        <Image
-            source={require('../../assets/ExpressoLogo.png')}
-            style={styles.headerIcon}
-        />
-        <TouchableOpacity onPress={() => {navigation.navigate('Cart')}}>
-          <Image source={require('../../assets/carticon.png')} style={styles.cartIcon}/>
-        </TouchableOpacity>
-      </View>
+      <Header></Header>
       <View style={styles.cartView}>
         <Text style={styles.cartTitle}>Cart</Text>
         <Cart receiveValue={receiveAmount}></Cart>
@@ -42,23 +35,6 @@ const CartScreen = ({navigation}) => {
 };
 
 const styles = StyleSheet.create({
-  navBar: {
-    marginBottom: 15,
-    marginTop: 8,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-  },
-  headerIcon: {
-    width: 200,
-    height: 50,
-    marginLeft: 15,
-  },
-  cartIcon: {
-    width: 40,
-    height: 40,
-    marginTop: 8,
-    marginRight: 10,
-  },
   cartView: {
     justifyContent: 'center',
     alignItems: 'center',
